module-4/requirements.txt
pandas
scikit-learn
joblib
matplotlib
seaborn
streamlit
python-magic
